Rewrite systemctl-commands.sh for portability
Reliably detect os and version.
Test systemd units are present before attempting to disable them.
Create /sbin/init as symlink iff necessary.

+++ platforms/docker/images/common-systemd/systemctl-commands.sh
@@ -1,47 +1,54 @@
 #!/bin/bash
 set -e -u -x
 export SYSTEMD_COLORS=0
+source /etc/os-release
+
 # Don't try to run a full system in the container
 systemctl enable basic.target
 systemctl set-default basic
-systemctl preset-all
-# Make sure we don't try to hijack gettys even if the user runs systemd in a
-# privileged container. We don't want it taking over the host TTYs.
-systemctl mask getty.target
-systemctl mask console-getty.service
-systemctl mask container-getty@.service
-systemctl mask getty-pre.target
-systemctl mask getty@.service
-systemctl mask serial-getty@.service
-# Don't mess with the clock
-systemctl mask time-sync.target
-# Unnecessary in a container:
-systemctl disable machines.target
-systemctl disable network-pre.target
-systemctl disable network.target
-systemctl disable remote-fs-pre.target
-systemctl disable remote-fs.target
-systemctl disable swap.target
-systemctl disable systemd-hwdb-update.service
-systemctl disable systemd-readahead-collect.service
-systemctl disable systemd-readahead-replay.service
-systemctl disable systemd-readahead-done.timer
-systemctl disable systemd-readahead-done.service
-# We don't need logins at all in a container
-systemctl mask systemd-logind.service
-# Don't try to set up fuse
-systemctl mask sys-fs-fuse-connections.mount
-# Prune unnecessary services
-systemctl disable proc-sys-fs-binfmt_misc.mount
-systemctl disable sys-kernel-config.mount
-systemctl disable sys-kernel-debug.mount
-systemctl disable systemd-ask-password-console.path
-systemctl disable systemd-ask-password-wall.path
+if [ $ID != "ubuntu" -o $VERSION_ID != "16.04" ]
+# preset-all can't cope with masked services in systemd 229
+then
+	systemctl preset-all
+fi
 
-source /etc/os-release
+installedunits=$( systemctl list-unit-files | head -n -1 | tail -n -1 | cut -f1 -d ' ')
+
+UNITS_TO_DISABLE=" machines.target network-pre.target network.target
+  remote-fs-pre.target remote-fs.target swap.target
+  systemd-hwdb-update.service systemd-readahead-collect.service
+  systemd-readahead-replay.service systemd-readahead-done.timer
+  systemd-readahead-done.service
+  proc-sys-fs-binfmt_misc.mount sys-kernel-config.mount
+  sys-kernel-debug.mount
+  systemd-ask-password-console.path systemd-ask-password-wall.path
+"
+
+UNITS_TO_MASK="
+	getty.target console-getty.service container-getty@.service
+	getty-pre.target getty@.service serial-getty@.service
+	time-sync.target
+	system-logind.service
+	sys-fs-fuse-connections.mount
+"
+
+for target in $UNITS_TO_DISABLE ;
+do
+	# need to check a unit is installed before trying to disable it
+	if [[ " ${installedunits[@]} " =~ " $target " ]] ; then
+		systemctl disable $target
+	fi
+done
+
+for target in $UNITS_TO_MASK ;
+do
+	systemctl mask $target
+done
+
 if [ $ID = "ubuntu" ]
 then
     systemctl disable apt-daily-upgrade.timer
     systemctl disable motd-news.timer
-    ln -s /bin/systemd /sbin/init
 fi
+
+[ -e /sbin/init ] || ln -s /bin/systemd /sbin/init
